demos/AI_OYD_Demo/requirements.txt
openai
streamlit
azure.storage.blob
azure.search.documents